Include Minecraft (Java Edition)
This is currently the #32 app on Flathub. Other than the Electron
Base app, and OpenJDK11 (SDK extension build), the Minecraft lanucher
is downloaded from mojang.com via extra-data, and downloads the game itself.

It appears there are a few advanced edge cases that have problems with the
Flatpak in https://github.com/flathub/com.mojang.Minecraft/issues, but generally
it works well as a Flatpak and the Flatpak is actively maintained.