Include Discord
Discord is currently the #2 download on Flathub.

The Flathub version of Discord is a redistribution of the (proprietary) Discord app
with permission, rather than an extra-data app:
https://github.com/flathub/com.discordapp.Discord/commit/587771e31a05a43565c03e8981e73eb86932be93

The Flatpak is actively maintained and well sandboxed (other than requiring X11 access).

Due to Flatpak limitations and maintainer decisions, some aspects of Discord either do not
work out of the box or do not work at all - see https://github.com/flathub/com.discordapp.Discord#readme -
However, this Flatpak still seems to be the best option for most people who want Discord
on Fedora.